Moved comment html and functions to Comment.js
